v1.3.0 | add APIManger support
