+++ GlowAPI/dependency-reduced-pom.xml
@@ -4,7 +4,7 @@
   <groupId>org.inventivetalent</groupId>
   <artifactId>glowapi</artifactId>
   <name>GlowAPI</name>
-  <version>1.2.1</version>
+  <version>1.3.0</version>
   <build>
     <sourceDirectory>src</sourceDirectory>
     <resources>
@@ -48,6 +48,7 @@
                   <include>org.mcstats.*</include>
                   <include>org.inventivetalent:glowapi**</include>
                   <include>org.inventivetalent:reflectionhelper**</include>
+                  <include>org.inventivetalent:apihelper**</include>
                 </includes>
               </artifactSet>
             </configuration>
@@ -102,7 +103,7 @@
     <dependency>
       <groupId>org.inventivetalent.packetlistener</groupId>
       <artifactId>api</artifactId>
-      <version>3.1.0</version>
+      <version>[3.1.0,)</version>
       <scope>compile</scope>
     </dependency>
   </dependencies>
