+++ GlowAPI/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.inventivetalent</groupId>
     <artifactId>glowapi</artifactId>
-    <version>1.2.1</version>
+    <version>1.3.0</version>
     <name>GlowAPI</name>
 
     <build>
@@ -53,6 +53,7 @@
                                     <include>org.mcstats.*</include>
                                     <include>org.inventivetalent:glowapi**</include>
                                     <include>org.inventivetalent:reflectionhelper**</include>
+                                    <include>org.inventivetalent:apihelper**</include>
                                 </includes>
                             </artifactSet>
                         </configuration>
@@ -93,12 +94,17 @@
         <dependency>
             <groupId>org.inventivetalent.packetlistener</groupId>
             <artifactId>api</artifactId>
-            <version>3.1.0</version>
+            <version>[3.1.0,)</version>
         </dependency>
         <dependency>
             <groupId>org.inventivetalent</groupId>
             <artifactId>reflectionhelper</artifactId>
-            <version>1.1.0</version>
+            <version>[1.2.3,)</version>
+        </dependency>
+        <dependency>
+            <groupId>org.inventivetalent</groupId>
+            <artifactId>apihelper</artifactId>
+            <version>[1.0.0,)</version>
         </dependency>
     </dependencies>
